requirements.txt
appdirs==1.4.4
arrow==0.15.7
asgiref==3.2.10
asn1crypto==1.3.0
attrs==19.3.0
base58==2.0.1
beautifulsoup4==4.9.1
binaryornot==0.4.4
bitarray==1.2.2
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coincurve==13.0.0
cookiecutter==1.7.2
cookiejar==0.0.3
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cytoolz==0.10.1
Django==3.0.8
django-extensions==3.0.2
djangorestframework==3.11.0
drf-yasg==1.17.1
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.0
ethereum==2.3.2
fake-useragent==0.1.11
future==0.18.2
hexbytes==0.2.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
inflection==0.5.0
ipfshttpclient==0.6.0.post1
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
lru-dict==1.1.6
lxml==4.5.1
MarkupSafe==1.1.1
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==20.4
pager==3.3
parse==1.15.0
parsimonious==0.8.1
parso==0.7.0
pbkdf2==1.3
poyo==0.5.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.5
ptpython==3.0.2
py-ecc==4.0.0
pycparser==2.20
pycryptodome==3.9.8
pyee==7.0.2
pyethash==0.1.27
Pygments==2.6.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.16.0
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
repoze.lru==0.7
requests==2.24.0
requests-html==0.10.0
rlp==1.2.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scrypt==0.8.15
secp256k1==0.13.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toolz==0.10.0
tqdm==4.47.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
varint==1.0.2
w3lib==1.22.0
wcwidth==0.2.5
web3==5.11.1
websockets==8.1
zipp==3.1.0
